Move sprint model and use jira models
